+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
-- Add `async` option to `start_link`
+- Add `Client.start` for non-linked processes.
+- Add `async` option to `start` and `start_link`.
 
 ## 0.1.3
 - `WebSockex.start_link` will no longer cause the calling process to exit on